+++ Sort/PyramidSort/heap.h
@@ -3,16 +3,16 @@
 
 #include "C:\algorithms\Lab2024\Sort\Qsort4\qsort.h"
 
-void siftDown       (int* array, int size, int index, int k);
+void siftDown       (int* array, size_t size, int index, int k);
 
-void Build          (int* array, int n, int k);
+void Build          (int* array, size_t n, int k);
 
-void heapSort       (int* array, int n, int k);
+void heapSort       (int* array, size_t n, int k);
 
-void heapSort2      (int* array, int n);
+void heapSort2      (int* array, size_t n);
 
-void heapSort3      (int* array, int n);
+void heapSort3      (int* array, size_t n);
 
-void heapSort4      (int* array, int n);
+void heapSort4      (int* array, size_t n);
 
 #endif // HEAP_H//
